+++ tests/Feature/Common/PublishableTraitTest.php
@@ -87,4 +87,36 @@ class PublishableTraitTest extends TestCase
 
         $this->assertCount(2, Page::getAllPublished());
     }
+
+    /** @test */
+    public function it_can_fetch_all_drafts_when_previewMode_is_active()
+    {
+        $this->asAdmin();
+        config()->set('chief.preview-mode', 'preview');
+
+        factory(Page::class)->create(['current_state' => PageState::DRAFT]);
+
+        $this->assertCount(1, Page::getAllPublished());
+    }
+
+    /** @test */
+    public function it_will_not_fetch_all_drafts_when_previewMode_is_active_but_no_admin_is_logged()
+    {
+        config()->set('chief.preview-mode', 'preview');
+
+        factory(Page::class)->create(['current_state' => PageState::DRAFT]);
+
+        $this->assertCount(0, Page::getAllPublished());
+    }
+
+    /** @test */
+    public function preview_mode_can_be_disabled()
+    {
+        $this->asAdmin();
+        config()->set('chief.preview-mode', null);
+
+        factory(Page::class)->create(['current_state' => PageState::DRAFT]);
+
+        $this->assertCount(0, Page::getAllPublished());
+    }
 }